Misc minor patches found during more testing and merging
Makefile.ocpi.common: build was missing CPPFLAGS in previous commit
Project.exports: AV merge in comments
components/file_read.rcc/file_read.xml: missed from previous commit, reduce size for unoc ocdp
hdl/platforms/modelsim/runSimExec.modelsim: fix verbose flag for modelsim script
runtime/util/property/src/OcpiUtilPValue.cxx: find one specific value after a generic one
scripts/install-liquid.sh: use newer installation script setup file properly
tools/ocpigen/src/parse.cxx: missed from previous commit (a warning).
runtime/application/src/ocpiexpr_main.cxx: add commented-out print for FP rounding mode issue
platforms/zed/zedsetup.sh: comment

+++ tools/ocpigen/src/parse.cxx
@@ -661,7 +661,7 @@ initImplPorts(ezxml_t xml, const char *element, PortCreate &a_create) {
       ezxml_set_attr_d(xml, "name", name.c_str());
     }
 #endif
-    if (!a_create(*this, x, NULL, nTotal == 1 ? -1 : ordinal, err))
+    if (!a_create(*this, x, NULL, nTotal == 1 ? -1 : (int)ordinal, err))
       return err;
   }
   return NULL;
